+++ app/backend/src/tests/test_admin.py
@@ -8,15 +8,13 @@ from sqlalchemy.sql import func
 from couchers.db import session_scope
 from couchers.models import (
     AccountDeletionToken,
-    Cluster,
     ContentReport,
     EventOccurrence,
     ModerationUserList,
-    Node,
     Reference,
     UserSession,
 )
-from couchers.proto import admin_pb2, auth_pb2, editor_pb2, events_pb2, references_pb2, reporting_pb2
+from couchers.proto import admin_pb2, auth_pb2, events_pb2, references_pb2, reporting_pb2
 from couchers.sql import couchers_select as select
 from couchers.utils import Timestamp_from_datetime, now, parse_date, timedelta
 from tests.test_communities import create_community
@@ -31,7 +29,6 @@ from tests.test_fixtures import (  # noqa
     mock_notification_email,
     push_collector,
     real_admin_session,
-    real_editor_session,
     references_session,
     reporting_session,
     testconfig,
@@ -381,178 +378,6 @@ def test_CreateApiKey(db, push_collector):
     push_collector.assert_user_has_single_matching(
         normal_user.id, title="An API key was created for your account", body="Details were sent to you via email."
     )
-
-
-VALID_GEOJSON_MULTIPOLYGON = """
-    {
-      "type": "MultiPolygon",
-      "coordinates":
-       [
-        [
-          [
-            [
-              -73.98114904754641,
-              40.7470284264813
-            ],
-            [
-              -73.98314135177611,
-              40.73416844413217
-            ],
-            [
-              -74.00538969848634,
-              40.734314779027144
-            ],
-            [
-              -74.00479214294432,
-              40.75027851544338
-            ],
-            [
-              -73.98114904754641,
-              40.7470284264813
-            ]
-          ]
-        ]
-      ]
-    }
-"""
-
-POINT_GEOJSON = """
-{ "type": "Point", "coordinates": [100.0, 0.0] }
-"""
-
-
-def test_CreateCommunity_invalid_geojson(db):
-    super_user, super_token = generate_user(is_superuser=True)
-    normal_user, normal_token = generate_user()
-    with real_editor_session(super_token) as api:
-        with pytest.raises(grpc.RpcError) as e:
-            api.CreateCommunity(
-                editor_pb2.CreateCommunityReq(
-                    name="test community",
-                    description="community for testing",
-                    admin_ids=[],
-                    geojson=POINT_GEOJSON,
-                )
-            )
-        assert e.value.code() == grpc.StatusCode.INVALID_ARGUMENT
-        assert e.value.details() == "GeoJson was not of type MultiPolygon."
-
-
-def test_CreateCommunity(db):
-    with session_scope() as session:
-        super_user, super_token = generate_user(is_superuser=True)
-        normal_user, normal_token = generate_user()
-        with real_editor_session(super_token) as api:
-            api.CreateCommunity(
-                editor_pb2.CreateCommunityReq(
-                    name="test community",
-                    description="community for testing",
-                    admin_ids=[],
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                )
-            )
-            community = session.execute(select(Cluster).where(Cluster.name == "test community")).scalar_one()
-            assert community.description == "community for testing"
-            assert community.slug == "test-community"
-
-
-def test_UpdateCommunity_invalid_geojson(db):
-    super_user, super_token = generate_user(is_superuser=True)
-
-    with session_scope() as session:
-        with real_editor_session(super_token) as api:
-            api.CreateCommunity(
-                editor_pb2.CreateCommunityReq(
-                    name="test community",
-                    description="community for testing",
-                    admin_ids=[],
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                )
-            )
-            community = session.execute(select(Cluster).where(Cluster.name == "test community")).scalar_one()
-
-            with pytest.raises(grpc.RpcError) as e:
-                api.UpdateCommunity(
-                    editor_pb2.UpdateCommunityReq(
-                        community_id=community.parent_node_id,
-                        name="test community 2",
-                        description="community for testing 2",
-                        geojson=POINT_GEOJSON,
-                    )
-                )
-            assert e.value.code() == grpc.StatusCode.INVALID_ARGUMENT
-            assert e.value.details() == "GeoJson was not of type MultiPolygon."
-
-
-def test_UpdateCommunity_invalid_id(db):
-    super_user, super_token = generate_user(is_superuser=True)
-
-    with real_editor_session(super_token) as api:
-        api.CreateCommunity(
-            editor_pb2.CreateCommunityReq(
-                name="test community",
-                description="community for testing",
-                admin_ids=[],
-                geojson=VALID_GEOJSON_MULTIPOLYGON,
-            )
-        )
-
-        with pytest.raises(grpc.RpcError) as e:
-            api.UpdateCommunity(
-                editor_pb2.UpdateCommunityReq(
-                    community_id=1000,
-                    name="test community 1000",
-                    description="community for testing 1000",
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                )
-            )
-        assert e.value.code() == grpc.StatusCode.NOT_FOUND
-        assert e.value.details() == "Community not found."
-
-
-def test_UpdateCommunity(db):
-    super_user, super_token = generate_user(is_superuser=True)
-
-    with session_scope() as session:
-        with real_editor_session(super_token) as api:
-            api.CreateCommunity(
-                editor_pb2.CreateCommunityReq(
-                    name="test community",
-                    description="community for testing",
-                    admin_ids=[],
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                )
-            )
-            community = session.execute(select(Cluster).where(Cluster.name == "test community")).scalar_one()
-            assert community.description == "community for testing"
-
-            api.CreateCommunity(
-                editor_pb2.CreateCommunityReq(
-                    name="test community 2",
-                    description="community for testing 2",
-                    admin_ids=[],
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                )
-            )
-            community_2 = session.execute(select(Cluster).where(Cluster.name == "test community 2")).scalar_one()
-
-            api.UpdateCommunity(
-                editor_pb2.UpdateCommunityReq(
-                    community_id=community.parent_node_id,
-                    name="test community 2",
-                    description="community for testing 2",
-                    geojson=VALID_GEOJSON_MULTIPOLYGON,
-                    parent_node_id=community_2.parent_node_id,
-                )
-            )
-            session.commit()
-
-            community_updated = session.execute(select(Cluster).where(Cluster.id == community.id)).scalar_one()
-            assert community_updated.description == "community for testing 2"
-            assert community_updated.slug == "test-community-2"
-
-            node_updated = session.execute(select(Node).where(Node.id == community_updated.parent_node_id)).scalar_one()
-            assert node_updated.parent_node_id == community_2.parent_node_id
 
 
 def test_GetChats(db):
