+++ app/backend/src/tests/test_email.py
@@ -19,7 +19,7 @@ from couchers.models import (
     User,
 )
 from couchers.notifications.notify import notify
-from couchers.proto import admin_pb2, api_pb2, events_pb2, notification_data_pb2, notifications_pb2
+from couchers.proto import api_pb2, editor_pb2, events_pb2, notification_data_pb2, notifications_pb2
 from couchers.sql import couchers_select as select
 from couchers.tasks import (
     enforce_community_memberships,
@@ -41,6 +41,7 @@ from tests.test_fixtures import (  # noqa
     process_jobs,
     push_collector,
     real_admin_session,
+    real_editor_session,
     session_scope,
     testconfig,
 )
@@ -440,8 +441,8 @@ def test_email_deleted_users_regression(db):
             api.RequestCommunityInvite(events_pb2.RequestCommunityInviteReq(event_id=event_id))
         assert mock.call_count == 1
 
-    with real_admin_session(super_token) as admin:
-        res = admin.ListEventCommunityInviteRequests(admin_pb2.ListEventCommunityInviteRequestsReq())
+    with real_editor_session(super_token) as editor:
+        res = editor.ListEventCommunityInviteRequests(editor_pb2.ListEventCommunityInviteRequestsReq())
         assert len(res.requests) == 1
         # this will count everyone
         assert res.requests[0].approx_users_to_notify == 5
@@ -450,15 +451,15 @@ def test_email_deleted_users_regression(db):
         session.execute(update(User).where(User.id == ban_user.id).values(is_banned=True))
         session.execute(update(User).where(User.id == delete_user.id).values(is_deleted=True))
 
-    with real_admin_session(super_token) as admin:
-        res = admin.ListEventCommunityInviteRequests(admin_pb2.ListEventCommunityInviteRequestsReq())
+    with real_editor_session(super_token) as editor:
+        res = editor.ListEventCommunityInviteRequests(editor_pb2.ListEventCommunityInviteRequestsReq())
         assert len(res.requests) == 1
         # should only notify creating_user, super_user and normal_user
         assert res.requests[0].approx_users_to_notify == 3
 
         with mock_notification_email() as mock:
-            admin.DecideEventCommunityInviteRequest(
-                admin_pb2.DecideEventCommunityInviteRequestReq(
+            editor.DecideEventCommunityInviteRequest(
+                editor_pb2.DecideEventCommunityInviteRequestReq(
                     event_community_invite_request_id=res.requests[0].event_community_invite_request_id,
                     approve=True,
                 )
